Parsing Methodology + Funciones de ayuda
Started thinking of ways to parse the NBT data, and output it to either JSON or an XML-like syntax. I think this would be a really nice way to edit NBT data, as it is a more visual way to edit it. And after thinking about it, NBT kind of is a markup-like way of holding data, so I think displaying it in that way may make it really nice to work with. It would be great to both be able to save write in it using an XML format, and get it's properties and values using object property dot notation.

I also added a function to check for Bedrock level dat headers at the beginning of the file. This was inspired from how prismarine-nbt checks for it.

Added a new function to display a number as plain binary ones and zeros. This is to help with my console logging to see what data is coming in an out of the different logic components as I write them.