Achievement Get! Endianness Conversion
Figured out how to flip integer endianness! The way I am doing it right now is fairly slow because it uses string manipulation, but I think this is what I will eventually have to do to read the Bedrock NBT format bytes correctly. This is awesome! Definitely on to something here.