Byte Endianness Article!
I think I may be getting closer again! I found a StackOverflow post that was going over how to convert a little endian file to a big endian file, and the answer described that the BYTES are rearranged, and now the bits. I have been trying the bits most of the time now, so I took a few wrong turns. I decided to make a short little article about this, just for fun, and to try and explain how it works to myself. Hopefully this is a correct find because I really like how my page turned out! I will definitely update it if the information is incorrect. I'm determined to get this down!!!