NBT Parser: ES Module Experimentation
Started out by breaking the library's functions up into individual modules! This feels kind of awkward at the moment, as it is mostly just the older code broken up a bit, but I think this will help me with compartmentalizing (a word? oof) how the logic works a bit more. Once I rewrite it, I may look into collapsing things back down again. I did like the simplicity of only having to load one file to get the library to work.

I also started testing a new NBT file, one that holds Structure Block data! It appears to be using some kind of compression, like zlib or gzip, I don't remember which. Because of that, I'll have to look into getting a compression library to decompress the data. I've found a few so far that may be of interest, zlib.js, and fflate.