Project Folder Restructure
Changes:
- Organized the project's files into nicer folders! The entry point for NBT-Parser is now just at /index.js, and the example page is now at /example/index.html instead.

I want to add support for use with JSDelivr, as it supports bundling ES Modules as you request them! I want to use this tool here, and allow NBT-Parser to be fetched as a minified file, with only one request. That kind of bundling is new to me, so I will start looking into how it works. I really like the setup that Jake Archibald uses for IDB-Keyval, and I want to provide something similar for NBT-Parser!
https://www.jsdelivr.com/github
https://github.com/jakearchibald/idb-keyval#all-bundles