Bundling with esbuild
Tried out a bundler for the first time! esbuild looks to be exactly what I was looking for. It is very performant, and it does a great job of making super small builds from modules, and it takes barely any time to do it. I will either look into writing NBT-Parser with modules and the building it with a build tool, or I will just write everything all in once place just like NBT.js does.

Feel free to check out esbuild, I highly recommend it! I also forgot to
mention that there's almost no setup for it, a very helpful note!
https://esbuild.github.io/