CommonJS Testing
Did some testing to see how to import from ES Modules into a CommonJS script. Pretty nice! It probably makes more sense to transpile with esbuild though, since awaiting the import function's return value isn't as nice as a standard require() call.
Edit: Since the package is in the ESM format in the package.json, you have to move the app.js entry point into the folder above the project in order for it to run as a CJS script.