Rewrite, now with writing support!!!
Rewrote the writing portion of the library. It works great! Now it natively supports compression cross-platform, out of the box.

Another change from the last commit I didn't mention yet is that the endianness and encoding are auto-detected now, so you don't even have to know them beforehand in order to open the file! Super neat. The only case where this could be a problem is if the compression format isn't supported by the Compression Streams API, as that only supports either "deflate" or "gzip".