Experimental Expand
Started working on a few functions to expand and simplify NBT data between JSON and back!
One super cool thing with this is that it will allow you to save plain JSON directly as raw NBT, without having to definte the types yourself. The one drawback with how this is currently implemented is that if you are converting from NBT to plain JSON, you lose the original inferrered types from the tag, which I am not quite fond of. They are estimated based on the value of the tag, which would be correctly stored, but could be different than what Minecraft may be expecting on it's own end.
I think a nice way to get around this issue would be to encode the tag value with the type inside of a string (like I am currently doing for BigInt values to help JavaScript out a little bit (not an intended byte pun, but I'll take it!)), instead of just the number, and then the original chosen type could be preserved with no issue.
For cases where you want to store JSON as NBT and switch between the two, it's not a problem to infer the types, since JSON only has the one number type, similarly to JavaScript, but it would be an issue when going from an original set of NBT data to JSON. Either way, this is super nice! (While also, kind a buggy atm)
I am planning on using this to parse Minecraft world data, as well as to store Flatlands JSON save data first as JSON, then as NBT, since it is much smaller in file size, especially when compressed with GZIP :D

This article, as well as the Minecraft Gamepedia NBT article helped out with getting the different ranges for number types:
https://www.tutorialspoint.com/cprogramming/c_data_types.htm