Individual Type Associations + ESM Antipattern Fix
Changes:
- Moved type declarations to be used directly by each module itself, rather than only from within the `declare module` syntax for Node. I want the type declarations to show up with regular local JS use in an editor too, not only with NPM usage. Ideally, a reference to a script file in a `git clone` copy of NBT Parser should also show the types in your editor, not exclusively with `npm install`. An example of this could be importing it from a local path, like `import * as NBT from "./scripts/nbt/src/index.js";`.
- Updated a few type declarations for the Reading and Writing parts of the library, just to make it more specific to the library's IO.
- Did some type checking using a `tsconfig.json` setup by using the `checkJS` option. I added a missing parameter to a `DataView().getInt16()` call, as it silently requires a parameter. I thought it defaults to `0`, as it never errored out, but TypeScript didn't like it, so I put `0` in to be more explicit.
- Added `deflate-raw` as one of the encoding options for the compression module! It appears to have recently been added in a new version of Chrome. It's not adopted in many places yet, but might as well allow it to be a parameter in terms of type checking, since it is valid there.
- I adjusted indentation and added more semicolons to the type declarations, but it messed with adding this commit, so I will add a second commit after this one to help show things that changed a little easier.
- Part of the edits have to do with the discovery of the ESM antipattern, which is something I was
stuck in: https://stackoverflow.com/questions/35541864/es6-export-default-function#comment58774983_35542310
- It's the relation between default exports and named exports, where you can't use the benefits of ESM loading if you are using both default and named exports,
because it will still force everything to load either way.

Here's why the syntax I want works for CommonJS modules, and not for ES Modules:
```js
// CommonJS
module.exports = {
  hey: function(){}
  yoo: function(){}
}

// ESM
export default {
  function hey(){}
  function yoo(){}
}
```

The import object for `module.exports` is just a plain object with properties, so you can destructure the same object without having to export it twice.
However, with ESM, you can't `import { hey } from "module"` if it's only a default export, because it's a module, and not a bare object that can be destructured.
And providing both named exports and a default object export is the antipattern for ESM, as it will force-load everything, so it can be loaded for use in the default export object. aaah!

I wish I could have `import NBT from "nbt-parser"`, as well as `import { read } from "nbt-parser"`, but looks like you can only do `import * as NBT from "nbt-parser"` instead
of the first option (in the context of using ESM).