Types Formatting
Adjusted the indentation from the `declare module` syntax removal, and added semicolons to the end of most lines, to be more *in line* with the rest of my syntax style (see what I did there :} ). This was part of the last commit, but I separated it so you can see what changed easier.