Separate Module Types
Yet another iteration of tying to figure out how to use `d.ts` files for my JS, ooof!
For this one, I split all of the declarations into separate modules files themselves. This appears to work for loading the library with local paths this time, but now not with NPM! Ideally I'd like to have the types in a separate folder, but that could be the main problem preventing this from working, not totally sure. While TypeScript is largely popular and documented, I'm finding it to be interesting that this is such a hard setup to get right. I'd think it wouldn't be this difficult to get a setup like I'm looking for working, or to at least find something like it that someone else has done.
But, it could also be the case that I'm just doing it wrong, haha. I don't really belive there can be a wrong way to implement it though, maybe a less ideal or intended way, but that doesn't mean impossible! I could be doing it wrong afterall, and that means I am getting even closer to doing it right at least :D

These links helped with this attempt:
https://stackoverflow.com/questions/60324367/typescript-declare-module-export-from-another-module
https://stackoverflow.com/questions/35065880/typescript-declare-module
https://stackoverflow.com/questions/43281741/how-to-use-paths-in-tsconfig-json