The Next Stage - Crazy rewrite!
Ok, so I almost have it all working, but not quite yet. There's gonna be a few more reworks, as I want to have it fully working and stable, both with the new features and API structure. Since so much is changing with the whole project at the moment, I'll officially go over all the changes once I have it working stable again.

Sorry there are so many crazy branches! With time, I definitely do want to get everything back into the one main branch.

*Edit later:
Yes, I'm coming back to go over this, haha

**Edit later later:
There was a demo for these changes that was fairly similar, and it had a better changelog for everything new, better than the one above this. I thought I'd keep that one, and add it below as a reference:

Crazy rewrite!

Added classes, and a ton of things.
Not quire working yet, so I'm gonna do yet another write with these changes *(this is the commit that this line references XD)*, just to reverse engineer the bugs before they get added, since the current stable implementation is, well, stable XD
I mostly did all of this yesterday, but I'm commiting it today since I want to kind of start fresh with it *(Aug 15, in regards to the original commit for this message. It was right before I added "The Reset", over on another branch though)*.

This library is essentially a complex combo of NBT.js, Prismarine-NBT, nbt-ts, and the NBT module in anvil-parser all combined. I like some parts from each of those, so I'm trying to make it exactly how I want it to work!

https://github.com/sjmulder/nbt-js
https://github.com/PrismarineJS/prismarine-nbt
https://github.dev/janispritzkau/nbt-ts
https://github.dev/Geoige/anvil-parser

A few other links that helped out a bit:
https://www.typescriptlang.org/docs/handbook/2/objects.html
https://stackoverflow.com/questions/56969950/optional-typed-immediately-destructured-parameters