Secure Public Value Property
Another option that this could take.

This kind of makes me want to rethink the extensions for the other classes, as this actually makes it fairly secure, and the logging actually looks fairly good in the console.

I may take a break the next few days, especially with AFC in just over a day or so, and come back with a fresh mind to decide, yet again, how the API should work.