Endian Default Docs
Oh yeah! AFC was EPIC!!! I was only 18 seconds away from my Rock 'n Roll PR, and that includes the climb up 6th Ave at the end. Wow, that one really hurts. I've been up that hill a few times now, but not recently. That's another level of dedication to not start walking. I was very close, I had to make a deal with myself that that was NOT going to happen, whatsoever. I think the rest of the race before that is what makes it so hard. Really happy that I got water along the way, that's really good for you :)