Test Folder Re-Addition
It's back! This was gone for a while, as I have made a ton of changes for the API over time, and I didn't want to confuse myself with older logic in there. Now I will be starting to work on the writing portion of the API, and addings docs and explainers for the public to use. Little steps!

Yeah so these files were around back in May, so yeah they've been scattered for a little while.

And this folder is more for checking feature behavior, rather than the test libraries some people use to check their APIs. Expect some of the code in here to randomly change, and it's probably not gonna be the neatest example of organization. Grain of salt, pls :P

This is a great resolution to loading files relatively with the `fs` module in Node! So happy I can run the test script from the root folder with no issue now.
https://stackoverflow.com/questions/46745014/alternative-for-dirname-in-node-js-when-using-es6-modules