JavaScript, Refactor?
I know, I'm going back and forth. This has been a huge learning journey, and that's my main goal.

I went back to JS, from the TypeScript rewrite. Now everything is 98% the same as before, the only different being that the TypeScript features are now defined using JSDoc, right inside of the vanilla JS! I really want to avoid a transpile step, unless it were completely necessary (not in this case, for me at least). All I essentially want from my use with TypeScript is variable type checking in my editor, and things like that. I also really like the automatic documentation that it adds to your library. It just pops up for the user with what options are available for them to use.

I guess, ideally, it would be wonderful to be able to write everything out in TypeScript, then transpile it to plain vanilla ESNext JavaScript, with JSDoc as the documentation side of things. I don't want to have to maintain a different version of the source which isn't directly what it used on the user's end though, which is my main eeerrrg with TypeScript. Comparing writing TypeScript vs JavaScript with JSDoc, it without a shadow of a doubt is better. But, if you count in the tooling that you have to manage after the fact, I don't like that you have to both manage a separate source code, and a separate user code (the transpiled version).

This test (how many am I going to do? I'm still not sure XD) tries to find an even more middle ground, which benefits from the static-typing nature of TypeScript in your editor, but without having to transpile anything before you can run it.

There are still a few issues that I didn't quite figure out yet, with the JSDoc setup. The main one is being able to define an index signature on a class, as I want to define a few optional properties on the `CompoundTag`. I think this may be solveable with a more-specific set of type defintions for those properties though, rather than just with string indexing.

The other issue I didn't figure out yet is also likely solveable with a better implementation on my part, and that's accessing the class which creates the object for an NBT tag. A better way to put it, right now I am trying to get a reference to `ByteTag` from a `new ByteTag()` instance, using `new ByteTag().constructor`. However, I haven't found a way to type-cast that to specifically target `ByteTag`, as that's ideally what would work best. From the time of writing that last sentence, I think I though of a possible solution, haha. It would also probably work to add a factory function like `isTag()`, but to get the constructor class of a tag instance, something like `getTagConstructor()`, maybe. Another option I read that someone recommended was essentially `Object.getPrototypeOf(new ByteTag()).constructor`, but I'm not sure if I like the verbosity of adding an extra call like that, only for a type declaration fix. But, it is a fix, so it's definitely an option.

Even with all of this work, sometimes it still feels like a complete mess. I think that means (it could be multiple/any of these) that I've worked on it too long today, it not actually a mess and I'm just confused, or it is a mess, haha. I think it could be helpful to take a break and work on something else though, that usually does miracles for when I come back later to it (Even this crazy update/change/test thing was one of those too).

Maybe I'll work some more on trying this statically-typed TS setup for vanilla JS with Num Text Component, I think I had something like that going before I moved this project to TypeScript. I'd also have to say that the in-editor linting has been massively cool, I definitely am a fan of that now, now that I've peeked behind the curtain of TypeScript, lol.

Cheers to experimenting!

Some migration docs help:
https://stackoverflow.com/questions/38877252/typescript-getting-a-class-constructor-of-the-object-at-runtime
https://stackoverflow.com/questions/70101584/how-to-cast-typescript-type-in-javascript-using-jsdoc
https://stackoverflow.com/questions/49836644/how-to-import-a-typedef-from-one-file-to-another-in-jsdoc-using-node-js
https://stackoverflow.com/questions/16017627/document-generic-type-parameters-in-jsdoc
https://stackoverflow.com/questions/51467835/how-to-use-jsdoc-to-document-an-es6-class-property
https://stackoverflow.com/questions/71274013/using-jsdoc-to-set-an-index-signature-on-an-es6-class