JS with Types. Not TS.
SUPER COOL! This is next level fancy.

Found out that you can fold block comments in VS Code using a keyboard shortcut! So, essentially now I have all of the magic from TypeScript, in vanilla JS, with *type declarations*, and I don't have to transpile anything!!! This is magical. I think I just snuck into the sweet spot I've been trying to get for some time now.

Fixes:
- Fixed function parameter optional destructuring definitions. It doesn't let you mark an object declaration as optional, so you can work around it by allowing it's type to be the object w/ properties, or plainly undefined. Works great!
- Added a temporary fix for the `constructor.TAG_BYTE` type issue.

Changes:
- Did some formatting for param declarations on some writer method definitions, so they are more similar to the rest of the definition comments. They also fold in VS Code better too.
- Removed extra class type param definitions, as they don't appear to work in JSDoc. Because of that, I had to add a type cast for `Object.fromEntries()` in `CompoundTag.valueOf()`.

For the first and second links below, I found an alternative to `@param` and `@returns`. It works really nice! I was going to use it for most function declarations, but I ended up not wanting to have inconsistent declaration structures.

The third link is kind of a workaround for the class param type casting, but it didn't quite apply to my situation.

The last two links have to do with the multiline comment folding, and a VS Code setting which toggles the folded code's background. I think it looks very clean to essentially just hide all of the docs when working on the implementation itself, when you don't worry about the IO of the function or whatever you are working on.

Resources:
https://github.com/microsoft/TypeScript/issues/48650#issuecomment-1120446341
https://www.typescriptlang.org/play?filetype=js#code/PQKhAEBcE8AcFMAEBvAPAFQHwApIHsBreAO0QC5F0BKRAXk0WwENzKAaAI1erofQF8QwAFDwAHrDwAnSIgBmAV2IBjSAEs8pJgBNtAMSarp0XIRI1kwxNcRT4kBVNKKV6zYh3bmnC1Zv-gYABnBzk5P39be0ctAG4IxH5hJOFQECsQRCg4JGQUFgoQqTViAHM2RC4KYgUAWw54KX5EjJEAG3tEfCJiAEZ4tIysmAQUSrw8DqZiAG0AXRbEIWEO2W6SACZ44WVNEI9dXroD-UN8KRN1vqp43eJ9zw3jzwMjC9MejZugA
https://stackoverflow.com/questions/38321681/javascript-jsdoc-how-to-document-new-es6-datatypes-like-map
https://linuxpip.org/vscode-collapse-expand/
https://stackoverflow.com/questions/60789437/how-do-i-change-the-background-color-of-the-folded-block-in-vscode