Here We Go!
This is the start date for when I started the rewrite that will appear in the next commit which follows this one. I worked on the changes for about a week and a half, off and on. I was doing a lot of experimentation and reworks overall, so I didn't keep commits for each step along the way, just the end result of my testing. TypeScript helped so much to get this going well! I really love it now. I think it will start to use it fully, across more of my projects now too. I also just recently found out about GitHub Packages, and I am hoping to use that as an alternative to npm! I want to try and keep everything all in once place, mainly the repo, but we'll see if that's doable, just like how the strict-JS with JSDoc wasn't quite up to the same level of magic that TypeScript provides. Maybe I'll give in and go to npm. We'll see!