Back to TS!
Ok, while it is doable to have a dual JS-TS kind of workflow, plain TypeScript is just too nice not to use on it's own! I really love how streamlined it is to write in TS. It's so nice, I think I may have to just fully jump right into it! I am really glad I tried the mixed workflow too though. That could definitely be a valid workflow for a smaller project that doesn't need to be written as type-safe. I want this one to be very tight though, so TS will be a great way to make that happen!

Ok, and down to the changes. It's not just bringing it back to TypeScript, I actually rewrote everything from scratch! I found how a few other JS NBT libraries made it work, and it inspired me even more on how to make mine work. Part of the drive behind the new library structure and being in TypeScript, is that I realized that it will be very cool to use my new TypeScript knowledge, and to write type definitions for in-game NBT files (like the Bedrock level.dat, or a player save data file), and be able to use them safely typed, in strict JavaScript or TypeScript (the mixed setup with JS, or either a regular TS workflow). Alright, here we go:

Additions:
- Created a new NBTData wrapper class, and it holds metadata about the contained NBT object, such as what endianness and compression type it used in the original ArrayBuffer. This makes it so you can read an NBT ArrayBuffer, detect its file format, and be able to save it directly back to that same format without having to manually keep a reference to what format it used. Now the library makes it visible to you!

Changes:
- Modified all tags that JavaScript equivalents to now use those, instead of implementing a custom primitive wrapper object for each one. Now only primitives with no JavaScript equivalents have custom primitive wrapper classes. Those include the Byte, Short, Int, and Float primitive types.
- This tag change makes it so the values for the tag types are the tags themselves, rather than being an object which holds "type" and "value" keys. To check their types, now the library uses things like typeof or instanceof to check the tag's type.
- This also greatly simplifies the reading and writing methods on each the Reader and Writer classes, as now they only have to deal with the primitive itself, and it doesn't have to generate the primitive wrapper object for it anymore.
- If what I think this is, is what it is, I think it makes the data more declaratively structured, if that's the right term for it. Now you can just define a TypeScript interface type for the data, and it will document the data like a plain object, since that's what it is! If my description of that didn't make enough sense, I will be working on demos for that fairly soon!
- Re-gzipped the example bigtest.nbt file, as it appears to have used a different gzip format back from a few years ago. It was making my read-to-write checks fail, even though the actual NBT output ArrayBuffer was the exact same. So now those checks will work the same, since both the import file and export file use the gzip format used in Chrome/NodeJS, whatever that may be. Not sure what flavor of gzip Notch used to compress the original file, but it has a bit different of a header, causing the tests to fail.

Removals:
- Removed the existing API JSDoc comments, I am going to write new ones from scratch to reflect all of the new changes I made in this new rewrite.