Rebranding: NBTify!
The official name has now arrived, NBTify! (like stringify) Couldn't quite decide on one so easily, and this seemed to be the best option. I wanted a name that is easy to remember, and simple to type. I was going to do something like WebNBT or Web-NBT (So it lines up with APIs like WebRTC, WebGPU, or Web Audio API), but I didn't like that it didn't start with NBT, and that there's already a "webNBT" project. While that one isn't JS-based, I didn't want to step on their toes for any naming confusions. So, went with making my own!

While shortlived, trying out GitHub Packages was really nice in the previous release, as you have everything tied into GitHub. It's as feature complete as npm for use in Node, but it turns out there isn't (yet) a way to load it from a CDN. I'm a little bit bummed about that, but I think it's for the best. So much of the industry lives off of npm, so I think it's time that I jump aboard too. So much more cool stuff I can do!

(Reflection time!)
Crazy to go through yet another leap in learning. So happy with how the library it turning out! It has given me so many more things to learn. TypeScript has been absolutely wonderful. Now that I'm at this stage, maybe I could look into making a browser NBT editor make with React!

Here's to learning!
*cheers*

Brandon :)