Off to npm! - Version patch fix
I messed up the first release to npm, and I ended up deleting it right after publishing it; That one was 1.0.1. Looks like there's a rule where you have to wait 24 hours to publish again after deleting a release. I tried to publish it again tonight, after the 24 hour mark, but looks like you can't publish the same version again, which is probably a good thing. So I'm incrementing the version up a patch number.

I need to get better at publishing things that are broken (and leaving them), then fixing it with a new patch, rather than replacing the broken version later. You can do that with Git repos, but that's probably a bonus more than anything. I'm gonna stop myself from doing that with npm, since that's not going to work well for people that may end up using NBTify.

Can't wait to say this, I have an npm package now!

(Listening to The Accolade (I), that's an epic song. Not related to any of the notes above, just thought I'd mention it :D )