NBTView
Thought of an idea to make the buffer reading and writing separated from the actual reader and writer classes, and to make the implementation defined on a custom DataView object instead. As I'm not totally sure how this will work yet, I am working on this on a separate branch, while it's in it's early stages. This is my second take at trying to get it to work. The first version was at a similar stage to this take, but going through all of the changes I made, all at once, felt risky and I didn't want to wreck the library with something that I wasn't sure if it would work completely yet.