TypedArray Primitive Handling
Thought of an idea! Rather than handling what values can go into the custom primitive classes, I can just use their equivalent TypedArray constructors to handle the incoming data. This makes the primitives IO behave the same as how it's equivalent TypedArray would handle it. So, if you pass an incorrect type into the constructor, the TypedArray will smartly coalesce the value input to a valid value that works for that primitive type. Say if you were to pass `135` to the Byte constructor, as a random example, it will coalesce the value to `-121` instead, which fits within it's value range. Super neat! The one thing that I'm not sure about is that this will simply figure out a value that will work, but not let you know that the input value wasn't proper. I could see this introducing unnoticed issues, which I'm not fond of, but then again, this is the behavior that the built in TypedArrays use, so I also think that gives merit to the behavior. I think I will end up going with that, since that's the nature of how JavaScript works too :)