Primitive ValueOf Inheritance
Figured out how to passthrough the type of the primitive's value directly to the `valueOf()` function! This also adds generic support, so you can specifically define what values are allowed for the Byte constructor. This is key for making type definitions for Minecraft level files, so this is a great find to finally have working.

(Coming in the next commit) I haven't figured out how to distinguish the classes between each other, since they all have the same shape. So, defining a type of `Byte<0 | 1 | 2 | 3>` won't error if you set `Short<3>` as the value, for example. So I need to figure out how to show the difference between each of the classes, since they are designed to be use non-interchangeably (that's the whole point of them vs the regular `number` type!).