Primitive Distinguishing!
Wow! This was a neat find, immediately commiting this find to the repo, asap. I been looking for solutions like these two for a few months now, so it's great to have both of them working! :O

After digging into the `.d.ts` files for the project, I remembered how when you define a `#privateField` member onto a class, that it adds the `#private` definition into the class, in the `.d.ts` file. Trying by adding a private field definition to the class, and trying the type safety between `new Byte()` and `new Short()`, it works when there is a private field present! Groundbreaking! So, then I thought, I don't really need a private field for anything, since I'm inheriting directly from the Number object for these classes, so I only need it for the type distinctions. With that in mind, I tried adding `declare` to the field, to make it `declare #privateField`, but turns out that's not allowed. Then I remembered that TypeScript has a compile-time similar feature, with just the `private` keyword. That worked too! So I can use `declare private privateField`, and having those on each of the classes will make them marked as different internally, awesome!!!

This article helped start that mind idea train:
https://stackoverflow.com/questions/64707399/typescript-generates-declaration-d-ts-file-with-private-field

After getting that working, I realized that the built in TypedArray classes must be doing something similar to make them non-compatible, so I looked into it by assigning a Uint8Array to an Int8Array-typed variable. Turns out they are distinguished apart using `[Symbol.toStringTag]`, which is really clever! So after this is commited, I will probably follow that path too, as that doesn't add anything specific for just typing.

I chose `PrimitiveValue` for the private member name just since that's it's internal name in the ECMA standard, or something like that, that I saw somewhere. That's how it's listed for primitive Number object wrappers in DevTools too, and I thought that would make sense to be the error that throws when it distinguishes between a correct and incorrect type.