StringTag Distinctions
Relating to the previous commit, this further addresses the issue of getting the custom primitive types to discern each from the other. This works great! It's how the default type definitions for the different TypedArrays work.

I found that I also had to specify that the return type is a *specific* string, rather than just the `string` type, as that will still have the same original issue as before, since I think `[Symbol.toStringTag]` will *always?* return a string. Type-wise though, I want TypeScript to know that it returns a *specific* string. Ha! can't believe this is working!!! So hyped to have finally gotten closer with this part. Having special classes for the Number types are very important to defining type-safe declarations for world save data structures. (In recent news to this also (to me at least), there's a new feature in the TypeScript beta called `satisfies`, and it will work perfect for saying that a level data structure must conform to the CompoundTag type. Making it extend the CompoundTag would allow it to have any property on it, and that's not what I want. I want to restrict what kinds of properties can be on the declaration itself, super complex. Man, TypeScript has been eye-opening. It's helped me learn JS a bit more too, which is also great)

This also makes it nicer to when you stringify the object too, if you're weird enough to manually get around the inherited `toString()` from Number's prototype.

Here's an example of that by the way:
```ts
// This will now show that this value and the type are incompatible, something I've been working towards for a while! Note that it works with both a different input value, *and* class name! Awesome! (Byte<0> is different than Short<0>, Byte<1> is different than Short<4>, it doesn't matter which parts, they aren't the same :D )

const thingo: Short<0> = new Byte(0);
console.log(Object.prototype.toString.call(thingo));
```

Getting a little refresh on `[Symbol.toStringTag]` helped out a bit too, here's the spec I went to for that:
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Symbol/toStringTag