ByteArrayTag TypedArray
Changed the ByteArrayTag type over to Int8Array instead of Uint8Array. I was looking over the NBT spec over on https://wiki.vg/NBT, and it mentions that the ByteArray type is signed. Uint8Array is the unsigned version of Int8Array, so I changed it over to use Int8Array, since that represents the signed version of the data. I originally went with Uint8Array because I thought that ByteArray was just used for embedded data inside of the NBT (I had been using Uint8Array a lot for other things, like the raw NBT data as a whole), but I decided that without taking the direct information in the spec into account. Glad I noticed it!