BooleanTag Type
Now that the custom primitive wrapper tag types have value type assertions, I made a boolean tag type! As it's described on the Minecraft Wiki for NBT (https://minecraft.fandom.com/wiki/NBT_format), it's not a tag type in the sense that it's part of the serialized NBT, but rather that it's when you know that a ByteTag can be represented as just true or false. It also accounts for the literal use of `true` and `false` on the JavaScript side, as I imagine it would be easier to use that at times, and for the NBT writer to convert that to a ByteTag with the value of `0` or `1` automatically.