Recursive Structures (Initial Support)
Trying to get the list and compound types working on the NBTView class! It's seeming difficult for those though, as it's hard to calculate the offset sizes of the different tags that don't have named defined lengths. You know how long the string tag is before you parse it, for example. For the compound tag though, you just parse it until it ends. To work around this, I have a few different options: I could go back to my full reader-writer model, I could mix the two, using both an NBTView object and reader writer objects that handle the recursive structures separately (my favored choice after writing all of this out), or calculating the sizes of the recursive structures before you try to read or write them. While that would work, I think it's more calculations that you shouldn't have to do, because you're going to have to iterate over all of the data twice, and I don't like that. I also thought that it doesn't really make sense to read part of an NBT structure at a specific byteOffset, as it will never be in the same offset position if the surrounding dog changes (data, accidentally typed that, I'm listening to Bober now XD), or if it's just happens to work out to be in the same spot. So, for example, it doesn't really make sense for me to provide a method to read an NBT structure as a ByteTag at `x` index, why would you want to do that? So, I may just keep the reading logic as part of an implementation detail of the library, as you won't really have indeividual tags in a file to read, nobody has (that I've heard of, haha) just a random file with arbitrary NBT values scattered throughout the bytes, you will just have a regular NBT structure to parse into an object.

I think I'm overthinking this a bit, but I learn a lot that way! I will definitely be taking a ton away from this NBTView test, including a small NBTView class that is similar to this one, but with some of the craziness removed. I think my main goal with abstracting the reader and writer implementations into a common NBTView class, is that they seemed fairly common, and I wanted to see if I could slim down the implementations for those. But, that also makes things unneccesarily complicated, aah!! Haha, coding is so fun :)

Mike Keneally Band - Dog! (Louie, Bober)

Second time listening to this album, first time all the way through to the end :D
I had to listen to it twice in a row just this morning, I already love it!